reorg message
