--- packages/wallet-sdk/src/core/message/ActionResult.ts
@@ -1,9 +1,0 @@
-import { SerializedEthereumRpcError } from ':core/error';
-
-export type ActionResult<T> =
-  | {
-      value: T;
-    }
-  | {
-      error: SerializedEthereumRpcError;
-    };
