+++ packages/wallet-sdk/src/core/message/index.ts
@@ -1,17 +1,5 @@
-import { UUID } from 'crypto';
-
-export type MessageID = UUID;
-
-export interface Message {
-  id: MessageID;
-  requestId?: MessageID;
-}
-
-export type MessageWithOptionalId<T extends Message> = Omit<T, 'id'> & Partial<Pick<T, 'id'>>;
-
-export function createMessage<T extends Message>(params: MessageWithOptionalId<T>): T {
-  return {
-    ...params,
-    id: params.id ?? crypto.randomUUID(),
-  } as T;
-}
+export * from './ConfigMessage';
+export * from './Message';
+export * from './RPCMessage';
+export * from './RPCRequest';
+export * from './RPCResponse';
