+++ packages/wallet-sdk/src/sign/scw/SCWSigner.ts
@@ -3,10 +3,8 @@ import { SCWKeyManager } from './SCWKeyManager';
 import { SCWStateManager } from './SCWStateManager';
 import { PopUpCommunicator } from ':core/communicator/PopUpCommunicator';
 import { standardErrors } from ':core/error';
-import { createMessage } from ':core/message';
+import { createMessage, RPCRequestMessage, RPCResponse, RPCResponseMessage } from ':core/message';
 import { Action, SupportedEthereumMethods, SwitchEthereumChainAction } from ':core/message/Action';
-import { RPCRequestMessage, RPCResponseMessage } from ':core/message/RPCMessage';
-import { RPCResponse } from ':core/message/RPCResponse';
 import { AppMetadata, RequestArguments } from ':core/provider/interface';
 import { AddressString } from ':core/type';
 import { ensureIntNumber } from ':core/type/util';
